Move register and login forms to their own components
